+++ html/test.html
@@ -4,7 +4,7 @@
   <title>test page for dialogue styling</title>
   <link rel="stylesheet" href="/css/experimental.css">
   <style>
-    .test {
+    .test-image {
       position: absolute;
       width: 645px;
       height: 133px;
@@ -23,20 +23,23 @@
         display: none;
       }
     }
+    img {
+      image-rendering: smooth;
+    }
   </style>
 </head>
 <body class="accurate nowrap">
   <article who="yoki" emotion="think">
     I don't think I can fall back <span class="break"></span>
     asleep after THAT.
-    <img src="/images/test/test1.png" class="test">
+    <img src="/images/test/test1.png" class="test-image">
   </article>
   <article>
     I was born and raised in the Castle<span class="break"></span>
     Town near here.<span class="break end"></span>
     Left sometime in my late youth to<span class="break"></span>
     travel the world.
-    <img src="/images/test/test2.png" class="test">
+    <img src="/images/test/test2.png" class="test-image">
   </article>
   <article class="wrap inaccurate">
     (You can disable "mix-blend-mode: difference" and/or "display: block" in devtools)
