+++ js/unbuildDEPdialogue.js
@@ -1,6 +1,7 @@
 /* for use when developing to make my life easier */
 const fs = require('fs');
-const { targetPage, readPage } = require("./buildDEPdialogue.js");
+const { targetPage } = require("./buildDEPdialogue.js");
+const { readPage } = require("./utils.js");
 
 const unbuild = async () => {
   const { document } = await readPage(targetPage);
